papaya_models/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
albumentations>=1.0.0
segmentation-models-pytorch>=0.3.0
timm>=0.6.0
wandb>=0.12.0
tensorboard>=2.7.0
tqdm>=4.62.0
Pillow>=8.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
pycocotools>=2.0.4
efficientnet-pytorch>=0.7.0
pytorch-lightning>=1.5.0
